+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 
-## 0.3.0 - pending
+## 0.3.1-SNAPSHOT - pending
 
-*	
+## 0.3.0 - 2020-05-17
+
+*	build: update dependencies
+*	add custom logging implementation (based on retrofit1 implementation)
+*	add user agent interceptor with builder settings
+*	add default okhttp3 cookieJar
+*	remove Futurefit 1 implementation
 
 ## 0.2.1 - 25/09/2019